continuing the refactoring #1, #5
