+++ src/NDArray.h
@@ -2,21 +2,21 @@
 #pragma once
 
 #include <algorithm>
+#include <stdexcept>
+
 #include <cstddef>
 #include <cassert>
 
-namespace old {
-  
+namespace wip {
+
 // The array storage
-template<size_t D, typename T>
+template<typename T>
 class NDArray
 {
 public:
 
-  static const size_t Dim = D;
-
   // Max size in any one dimension of ~1e9
-  static const size_t MaxSize = 1u << 30;
+  static const int64_t MaxSize = 1 << 30;
 
   typedef T value_type;
 
@@ -25,37 +25,32 @@ public:
   typedef T& reference;
 
   // RW iterator over one dimension (O) of an n-D array given an index
-  template<size_t O>
   class ConstIterator
   {
   public:
 
-    static const size_t Orient = O;
-
-    ConstIterator(const NDArray<D, T>& a, size_t* idx) : m_a(a)
+    ConstIterator(const NDArray<T>& a, int64_t orient, const std::vector<int64_t>& idx) : m_a(a), m_orient(orient), m_idx(idx)
     {
-      // copy indices
-      std::copy(idx, idx + D, m_idx);
       // set index of orientation dimension to zero
-      m_idx[O] = 0;
+      m_idx[orient] = 0;
     }
 
     ~ConstIterator() { }
 
-    const size_t* idx() const
+    const std::vector<int64_t>& idx() const
     {
       return m_idx;
     }
 
     ConstIterator& operator++()
     {
-      ++m_idx[Orient];
+      ++m_idx[m_orient];
       return *this;
     }
 
     bool end() const
     {
-      return m_idx[Orient] >= m_a.size(Orient);
+      return (size_t)m_idx[m_orient] >= m_a.size(m_orient);
     }
 
     const_reference operator*() const
@@ -65,78 +60,76 @@ public:
 
   private:
     const NDArray& m_a;
-    size_t m_idx[D];
+    int64_t m_orient;
+    std::vector<int64_t> m_idx;
   };
 
-  // RO iterator over one dimension (O) of an n-D array given an index
-  template<size_t O>
-  class Iterator
+//   // RO iterator over one dimension (O) of an n-D array given an index
+//   template<size_t O>
+//   class Iterator
+//   {
+//   public:
+
+//     static const size_t Orient = O;
+
+//     Iterator(NDArray<D, T>& a, size_t* idx) : m_a(a)
+//     {
+//       // copy indices
+//       std::copy(idx, idx + D, m_idx);
+//       // set index of orientation dimension to zero
+//       m_idx[Orient] = 0;
+//     }
+
+//     ~Iterator() { }
+
+//     const size_t* idx() const
+//     {
+//       return m_idx;
+//     }
+
+//     Iterator& operator++()
+//     {
+//       ++m_idx[Orient];
+//       return *this;
+//     }
+
+//     bool end() const
+//     {
+//       return m_idx[Orient] >= m_a.size(Orient);
+//     }
+
+//     const_reference operator*() const
+//     {
+//       return m_a[m_idx];
+//     }
+
+//     reference operator*()
+//     {
+//       return m_a[m_idx];
+//     }
+
+//   private:
+//     NDArray& m_a;
+//     size_t m_idx[D];
+//   };
+
+  NDArray() : m_dim(0), m_sizes(), m_storageSize(0), m_data(0), m_owned(true)
   {
-  public:
-
-    static const size_t Orient = O;
-
-    Iterator(NDArray<D, T>& a, size_t* idx) : m_a(a)
-    {
-      // copy indices
-      std::copy(idx, idx + D, m_idx);
-      // set index of orientation dimension to zero
-      m_idx[Orient] = 0;
-    }
-
-    ~Iterator() { }
-
-    const size_t* idx() const
-    {
-      return m_idx;
-    }
-
-    Iterator& operator++()
-    {
-      ++m_idx[Orient];
-      return *this;
-    }
-
-    bool end() const
-    {
-      return m_idx[Orient] >= m_a.size(Orient);
-    }
-
-    const_reference operator*() const
-    {
-      return m_a[m_idx];
-    }
-
-    reference operator*()
-    {
-      return m_a[m_idx];
-    }
-
-  private:
-    NDArray& m_a;
-    size_t m_idx[D];
-  };
-
-  NDArray() : m_storageSize(0), m_data(0), m_owned(true)
-  {
-    m_sizes[0] = 0;
-    m_sizesl[0] = 0;
   }
 
-  explicit NDArray(const size_t* sizes) : m_storageSize(0), m_data(0), m_owned(true)
+  explicit NDArray(const std::vector<int64_t>& sizes) : m_dim(sizes.size()), m_sizes(sizes), m_storageSize(0), m_data(0), m_owned(true)
   {
     resize(sizes);
   }
-  
+
   // Construct with storage managed by some other object
-  // TODO disable stuff like resize
-  NDArray(int64_t* sizes, T* const storage)
+  NDArray(const std::vector<int64_t>& sizes, T* const storage)
+    : m_dim(sizes.size()), m_sizes(sizes)
   {
-    std::copy(sizes, sizes + Dim, m_sizes);
-    std::copy(sizes, sizes + Dim, m_sizesl);
+    assert(m_sizes.size());
     m_storageSize = sizes[0];
     assert(m_storageSize < MaxSize);
-    for (size_t i = 1; i < Dim; ++i)
+    for (size_t i = 1; i < m_dim; ++i)
     {
       assert(sizes[i] < MaxSize);
       m_storageSize *= sizes[i];
@@ -148,12 +141,12 @@ public:
 
   // Disallow copy
   NDArray(const NDArray&) = delete;
+  NDArray& operator=(const NDArray&) = delete;
 
   // But allow move
   NDArray(NDArray&& a)
   {
-    std::copy(a.m_sizes, a.m_sizes + Dim, m_sizes);
-    std::copy(a.m_sizesl, a.m_sizesl + Dim, m_sizesl);
+    m_sizes.swap(a.m_sizes);
     m_storageSize = a.m_storageSize;
     m_data = a.m_data;
     m_owned = a.m_owned;
@@ -166,21 +159,20 @@ public:
       deallocate(m_data);
   }
 
+  size_t dim() const
+  {
+    return m_dim;
+  }
+
   size_t size(size_t dim) const
   {
-    assert(dim < Dim);
+    assert(dim < m_dim);
     return m_sizes[dim];
   }
 
-  const size_t* sizes() const
+  const std::vector<int64_t>&  sizes() const
   {
     return m_sizes;
-  }
-
-  // python integration (TODO decide and remove one)
-  const long* sizesl() const
-  {
-    return m_sizesl;
   }
 
   size_t storageSize() const
@@ -193,17 +185,17 @@ public:
     return m_data;
   }
 
-  void resize(const size_t* sizes)
+  void resize(const std::vector<int64_t>& sizes)
   {
     if (m_owned)
     {
+      m_dim = sizes.size();
       size_t oldStorageSize = m_storageSize;
 
-      std::copy(sizes, sizes + Dim, m_sizes);
-      std::copy(sizes, sizes + Dim, m_sizesl);
+      m_sizes = sizes;
       m_storageSize = sizes[0];
       assert(m_storageSize < MaxSize);
-      for (size_t i = 1; i < Dim; ++i)
+      for (size_t i = 1; i < m_dim; ++i)
       {
         assert(sizes[i] < MaxSize);
         m_storageSize *= sizes[i];
@@ -216,6 +208,10 @@ public:
         m_data = allocate(m_storageSize);
       }
     }
+    else
+    {
+      throw std::runtime_error("resizing not permitted when memory is not owned");
+    }
   }
 
   void assign(T val) const
@@ -223,12 +219,22 @@ public:
     std::fill(m_data, m_data + m_storageSize, val);
   }
 
-  reference operator[](const size_t* index)
+  reference operator[](const std::vector<int64_t>& index)
   {
     return m_data[offset(index)];
   }
 
-  const_reference operator[](const size_t* index) const
+  const_reference operator[](const std::vector<int64_t>& index) const
+  {
+    return m_data[offset(index)];
+  }
+
+  reference operator[](const std::vector<int64_t*>& index)
+  {
+    return m_data[offset(index)];
+  }
+
+  const_reference operator[](const std::vector<int64_t*>& index) const
   {
     return m_data[offset(index)];
   }
@@ -243,7 +249,7 @@ public:
     return m_data + m_storageSize;
   }
 
-  // relinqish ownership
+  // relinqish ownership (caller must delete)
   void release()
   {
     m_data = nullptr;
@@ -251,12 +257,12 @@ public:
 
 private:
 
-  size_t offset(const size_t* const idx) const
+  size_t offset(const std::vector<int64_t>& idx) const
   {
-    // TODO this is pretty horrible, but it works. Template it?
+    // TODO this is pretty horrible, but it works.
     size_t ret = 0;
     size_t mult = m_storageSize;
-    for (size_t i = 0; i < Dim; ++i)
+    for (size_t i = 0; i < m_dim; ++i)
     {
       mult /= m_sizes[i];
       ret += mult * idx[i];
@@ -264,30 +270,37 @@ private:
     return ret;
   }
 
+  size_t offset(const std::vector<int64_t*>& idx) const
+  {
+    // TODO this is pretty horrible, but it works.
+    size_t ret = 0;
+    size_t mult = m_storageSize;
+    for (size_t i = 0; i < m_dim; ++i)
+    {
+      mult /= m_sizes[i];
+      ret += mult * *idx[i];
+    }
+    return ret;
+  }
+
   T* allocate(size_t size) const
   {
     return new T[size];
-    //return (T*)PyMem_Malloc(size * sizeof(T));
   }
 
   void deallocate(T* p) const
   {
     delete [] p;
-    //PyMem_Free(p);
   }
-
 
 private:
 
-  size_t m_sizes[D];
-  long m_sizesl[D];
+  size_t m_dim;
+  std::vector<int64_t> m_sizes;
   size_t m_storageSize;
   T* m_data;
   bool m_owned;
 };
 
-// 0d & 1d not implemented for obvious reasons
-template<typename T> class NDArray<0, T>;
-template<typename T> class NDArray<1, T>;
 
 }
