+++ src/NDArrayUtils.cpp
@@ -1,5 +1,5 @@
 
-#include "NDArrayUtils2.h"
+#include "NDArrayUtils.h"
 
 int32_t maxAbsElement(const std::vector<int32_t>& r)
 {
